[DK-82] refactor: 패키지 네이밍 변경
