+++ src/main/java/com/kdt/team04/domain/matches/proposal/entity/MatchProposalStatus.java
@@ -1,4 +1,4 @@
-package com.kdt.team04.domain.matches.request.entity;
+package com.kdt.team04.domain.matches.proposal.entity;
 
 public enum MatchProposalStatus {
 	WAITING, APPROVED
