+++ src/main/java/com/kdt/team04/domain/matches/proposal/repository/MatchProposalRepository.java
@@ -1,8 +1,8 @@
-package com.kdt.team04.domain.matches.request.repository;
+package com.kdt.team04.domain.matches.proposal.repository;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 
-import com.kdt.team04.domain.matches.request.entity.MatchProposal;
+import com.kdt.team04.domain.matches.proposal.entity.MatchProposal;
 
 public interface MatchProposalRepository extends JpaRepository<MatchProposal, Long> {
 }
